+++ tests/row_generation/test_PlainHuntGenerator.py
@@ -12,30 +12,30 @@ class PlainHuntGeneratorTests(TestCase):
         generator = PlainHuntGenerator(stage)
 
         initial_rows = gen_rows(generator, 4)
-        self.assertEqual([[2, 1, 3, 4],
-                          [2, 3, 1, 4],
-                          [3, 2, 1, 4],
-                          [3, 1, 2, 4]],
+        self.assertEqual([[2, 1, 3],
+                          [2, 3, 1],
+                          [3, 2, 1],
+                          [3, 1, 2]],
                          initial_rows)
         generator.reset()
         # Starts again after rounds
         second_rows = gen_rows(generator, 4)
-        self.assertEqual([[2, 1, 3, 4],
-                          [2, 3, 1, 4],
-                          [3, 2, 1, 4],
-                          [3, 1, 2, 4]],
+        self.assertEqual([[2, 1, 3],
+                          [2, 3, 1],
+                          [3, 2, 1],
+                          [3, 1, 2]],
                          second_rows)
 
     def test_singles(self):
         stage = 3
         generator = PlainHuntGenerator(stage)
         rows = gen_rows(generator, 6)
-        self.assertEqual([[2, 1, 3, 4],
-                          [2, 3, 1, 4],
-                          [3, 2, 1, 4],
-                          [3, 1, 2, 4],
-                          [1, 3, 2, 4],
-                          [1, 2, 3, 4]],
+        self.assertEqual([[2, 1, 3],
+                          [2, 3, 1],
+                          [3, 2, 1],
+                          [3, 1, 2],
+                          [1, 3, 2],
+                          [1, 2, 3]],
                          rows)
 
     def test_minimus(self):
@@ -58,9 +58,9 @@ class PlainHuntGeneratorTests(TestCase):
         generator = PlainHuntGenerator(stage)
 
         rows = gen_rows(generator, 3)
-        self.assertEqual([[2, 1, 4, 3, 6, 5, 8, 7, 10, 9, 11, 12],
-                          [2, 4, 1, 6, 3, 8, 5, 10, 7, 11, 9, 12],
-                          [4, 2, 6, 1, 8, 3, 10, 5, 11, 7, 9, 12]],
+        self.assertEqual([[2, 1, 4, 3, 6, 5, 8, 7, 10, 9, 11],
+                          [2, 4, 1, 6, 3, 8, 5, 10, 7, 11, 9],
+                          [4, 2, 6, 1, 8, 3, 10, 5, 11, 7, 9]],
                          rows)
 
     def test_maximus(self):