+++ wheatley/row_generation/helpers.py
@@ -5,8 +5,9 @@ from typing import List
 import itertools
 import re
 
+from wheatley.bell import Bell, BELL_NAMES
+
 _CROSS_PN = []
-_LOOKUP_NAME = "!1234567890ET"
 
 STAGES = {
     "singles": 3,
@@ -18,7 +19,11 @@ STAGES = {
     "caters": 9,
     "royal": 10,
     "cinques": 11,
-    "maximus": 12
+    "maximus": 12,
+    "septuples": 13,
+    "fourteen": 14,
+    "sextuples": 15,
+    "sixteen": 16,
 }
 
 
@@ -50,14 +55,19 @@ def convert_pn(pn_str: str, expect_symmetric=False) -> List[List[int]]:
 def convert_bell_string(bell: str) -> int:
     """ Convert a single-char string representing a bell into an integer. """
     try:
-        return _LOOKUP_NAME.index(bell)
+        return BELL_NAMES.index(bell) + 1
     except ValueError as e:
         raise ValueError(f"'{bell}' is not known bell symbol") from e
 
 
 def convert_to_bell_string(bell: int) -> str:
     """ Convert an integer into the equivalent bell name. """
-    if bell <= 0 or bell >= len(_LOOKUP_NAME):
+    if bell <= 0 or bell >= len(BELL_NAMES) + 1:
         raise ValueError(f"'{bell}' is not known bell number")
 
-    return _LOOKUP_NAME[bell]
+    return BELL_NAMES[bell - 1]
+
+
+def rounds(number_of_bells) -> List[Bell]:
+    """ Generate rounds on the given number of bells. """
+    return [Bell.from_number(i) for i in range(1, number_of_bells + 1)]